Fix apps to use new sparse accumulator and move old one to OLD
